+++ piflow-web/src/main/java/cn/cnic/component/visual/service/impl/GraphTemplateServiceImpl.java
@@ -138,7 +138,7 @@ public class GraphTemplateServiceImpl implements GraphTemplateService {
     @Override
     public ResponseResult delGraphTemplate(GraphTemplate graphTemplate) {
         //判断图表模板中是否有使用该连接
-        Long count = getGraphConfCunt(graphTemplate.getId());
+        Integer count = getGraphConfCunt(graphTemplate.getId());
         if(count  ==  0){
             GraphTemplate graphTemplateDelete = graphTemplateMapper.selectById(graphTemplate.getId());
             Integer assoId = graphTemplateDelete.getExcelAssoId();
@@ -180,10 +180,10 @@ public class GraphTemplateServiceImpl implements GraphTemplateService {
         return graphTemplateMapper.selectById(id);
     }
 
-    public long getGraphConfCunt(int id){
+    public Integer getGraphConfCunt(int id){
         QueryWrapper<GraphConf> wrapper = new QueryWrapper<>();
         wrapper.eq("graph_template_id",id);
-        Long count = graphConfMapper.selectCount(wrapper);
+        Integer count = graphConfMapper.selectCount(wrapper);
         return count;
     }
 }
